requirements.txt

aiogram
python-dotenv
setuptools
typer>=0.9.0
pydantic
pydantic-settings
pydantic-extra-types
PyYAML>=6.0
ruamel.yaml
aioredis
redis
cryptography>=41.0.0
pip-audit

SQLAlchemy[asyncio] 
alembic

aiosqlite                
psycopg[binary]   
# asyncpg>=0.27.0           
aiomysql           

loguru
parsedatetime
apscheduler

typer[all] 

aiohttp
requests

packaging

cachetools
rich

psutil
uvicorn[standard]

# Web panel dependencies
fastapi
starlette
itsdangerous
jinja2
PyJWT>=2.8.0
python-multipart
websockets

pytest
pytest-asyncio
pylint
sphinx
coverage
black
isort
mypy